+++ autoloader.php
@@ -9,8 +9,9 @@ include __DIR__ . '/src/Darya/Common/Autoloader.php';
 
 use Darya\Common\Autoloader;
 
+// Base path wherever it's included from, absolute path to the framework
 $autoloader = new Autoloader(realpath('./'), array(
-	'Darya' => 'vendor/darya/framework/src'
+	'Darya' => realpath(__DIR__ . '/src')
 ));
 
 $autoloader->register();