+++ src/Darya/Database/Connection.php
@@ -28,7 +28,7 @@ interface Connection {
 	public function disconnect();
 	
 	/**
-	 * Translate a storage query to a query for this connection.
+	 * Translate a storage query to a database query for this connection.
 	 * 
 	 * @param StorageQuery $storageQuery
 	 * @return \Darya\Database\Query