+++ src/Darya/Storage/InMemory.php
@@ -5,6 +5,9 @@ use Darya\Storage\Aggregational;
 use Darya\Storage\Filterer;
 use Darya\Storage\Readable;
 use Darya\Storage\Modifiable;
+use Darya\Storage\Query;
+use Darya\Storage\Queryable;
+use Darya\Storage\Result;
 use Darya\Storage\Searchable;
 use Darya\Storage\Sorter;
 
@@ -15,7 +18,7 @@ use Darya\Storage\Sorter;
  * 
  * @author Chris Andrew <chris@hexus.io>
  */
-class InMemory implements Readable, Modifiable, Searchable, Aggregational {
+class InMemory implements Readable, Modifiable, Searchable, Aggregational, Queryable {
 	
 	/**
 	 * The in-memory data.
@@ -102,6 +105,10 @@ class InMemory implements Readable, Modifiable, Searchable, Aggregational {
 	 */
 	public function listing($resource, $fields, array $filter = array(), $order = array(), $limit = 0, $offset = 0) {
 		$data = $this->read($resource, $filter, $order, $limit, $offset);
+		
+		if (empty($fields) || $fields === '*') {
+			return $data;
+		}
 		
 		$fields = (array) $fields;
 		
@@ -259,6 +266,63 @@ class InMemory implements Readable, Modifiable, Searchable, Aggregational {
 		}
 		
 		return array_unique($list);
+	}
+	
+	/**
+	 * Execute the given query.
+	 * 
+	 * @param Query $query
+	 * @return Result
+	 */
+	public function execute(Query $query)
+	{
+		$data = array();
+		$info = array();
+		
+		switch ($query->type) {
+			case Query::CREATE:
+				$this->create($query->resource, $query->data);
+				break;
+			case Query::READ:
+				$data = $this->listing(
+					$query->resource,
+					$query->fields,
+					$query->filter,
+					$query->order,
+					$query->limit,
+					$query->offset
+				);
+				break;
+			case Query:UPDATE:
+				$info['affected'] = $this->update(
+					$query->resource,
+					$query->data,
+					$query->filter,
+					$query->limit
+				);
+				break;
+			case Query::DELETE:
+				$this->delete(
+					$query->resource,
+					$query->filter,
+					$query->limit
+				);
+				break;
+		}
+		
+		return new Result($query, $data, $info);
+	}
+	
+	/**
+	 * Open a query on the given resource.
+	 * 
+	 * @param string       $resource
+	 * @param array|string $fields   [optional]
+	 * @return Query\Builder
+	 */
+	public function query($resource, $fields = array())
+	{
+		return new Query\Builder(new Query($resource, (array) $fields), $this);
 	}
 	
 	/**